+++ src/main/java/com/project/ohflix/domain/mylist/MyListController.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.project.ohflix._core.utils.ApiUtil;
+import com.project.ohflix.domain.content.Content;
 import com.project.ohflix.domain.profileIcon.ProfileIcon;
 import com.project.ohflix.domain.user.SessionUser;
 import com.project.ohflix.domain.user.UserService;
@@ -33,90 +34,26 @@ public class MyListController {
     private final ObjectMapper objectMapper;
 
 
+
     @PostMapping("/api/my-favorite-list")
     @ResponseBody
     public ResponseEntity<List<MyListResponse.ContentDTO>> sendRequestToOpenAI(@RequestBody MyListRequest.OpenAIRequest openAIRequest) {
-        return processOpenAIRequest(openAIRequest);
+        return myListService.processOpenAIRequest(openAIRequest);
     }
 
-    //restapi 처리할거
-    public ResponseEntity<List<MyListResponse.ContentDTO>> processOpenAIRequest(MyListRequest.OpenAIRequest openAIRequest) {
-        try {
-            // 메시지 생성
-            ObjectNode messageNode = objectMapper.createObjectNode();
-            messageNode.put("role", "user");
-            messageNode.put("content", openAIRequest.getMessage());
-
-            // 요청 본문 생성
-            ObjectNode requestBodyNode = objectMapper.createObjectNode();
-            requestBodyNode.put("model", openAIRequest.getModel());
-            requestBodyNode.set("messages", objectMapper.createArrayNode().add(messageNode));
-
-            String requestBody = objectMapper.writeValueAsString(requestBodyNode);
-
-            // 요청 본문 출력 (디버깅용)
-            System.out.println("Request Body: " + requestBody);
-
-            // RestTemplate 인스턴스 생성
-            RestTemplate restTemplate = new RestTemplate();
-
-            // 헤더 설정
-            HttpHeaders headers = new HttpHeaders();
-            headers.setContentType(MediaType.APPLICATION_JSON);
-            headers.setBearerAuth("");
-
-            // 요청 엔터티 생성
-            HttpEntity<String> entity = new HttpEntity<>(requestBody, headers);
-
-            // OpenAI API 호출
-            ResponseEntity<String> response = restTemplate.exchange(
-                    OPENAI_API_URL,
-                    HttpMethod.POST,
-                    entity,
-                    String.class
-            );
-
-            // 응답 본문 출력 (디버깅용)
-            System.out.println("Response Body: " + response.getBody());
-
-            // 응답 본문 파싱
-            String responseBody = response.getBody();
-            List<MyListResponse.ContentDTO> recommendedMovies = new ArrayList<>();
-            if (responseBody != null) {
-                // 응답 본문이 JSON 형식이 아닌 경우 텍스트 형식으로 처리
-                if (responseBody.startsWith("{")) {
-                    // JSON 형식인 경우
-                    recommendedMovies = myListService.parseRecommendedMovies(responseBody);
-                } else {
-                    // 텍스트 형식인 경우
-                    recommendedMovies = myListService.parseRecommendedMoviesFromText(responseBody);
-                }
-            }
-
-            // 추천 영화 목록 출력
-            recommendedMovies.forEach(movie -> System.out.println("Recommended Movie: " + movie));
-
-            return ResponseEntity.ok(recommendedMovies);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-        }
-    }
 
 
     @GetMapping("/api/my-favorite-list")
     public String getMyFavList(HttpServletRequest request) {
         SessionUser sessionUser = (SessionUser) session.getAttribute("sessionUser");
         MyListResponse.MyListDTO respDTO = myListService.findMyListById(sessionUser.getId());
-        String userMessage =myListService.getOpenAi();
-        MyListRequest.OpenAIRequest openAIRequest = new MyListRequest.OpenAIRequest("gpt-3.5-turbo-0125", userMessage);
-        List<MyListResponse.ContentDTO> recommendedMovies = processOpenAIRequest(openAIRequest).getBody();
-        //여기서부터하기!!!
-
+        List<Content> openAIRequest =myListService.getOpenAi();
+        System.out.println("openAIRequest = " + openAIRequest);
         request.setAttribute("MyListDTO", respDTO);
         request.setAttribute("openAIRequest", openAIRequest);
         return "mylist/my-favorite-list";
     }
+
 
     @PostMapping("/api/users/{contentId}/favorite")
     public ResponseEntity<?> addFavorite(@PathVariable int contentId, MyListRequest.AddFavoriteDTO reqDTO) {
